requirements.txt
streamlit==1.53.0
pandas==2.1.4
SQLAlchemy==2.0.46
langchain-openai==0.1.8
langchain-core==0.2.38
langchain==0.2.6
langchain-community==0.2.6
langchain-text-splitters==0.2.4
langgraph==0.1.19
langsmith==0.1.77
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
python-dotenv==1.2.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1